Split out PDU items to their own module.
